refactor(home): create base tab bar controller
