+++ GymApp/Coordinator/TabNavigationCoordinator.swift
@@ -15,36 +15,15 @@ enum Tab {
 
 class TabNavigationCoordinator: NSObject, TabBasedCoordinator {
 
-    internal lazy var rootViewController: UIViewController? = UITabBarController()
+    internal lazy var rootViewController: UIViewController? = BaseTabBarController()
     var childCoordinators: [StackBasedCoordinator] = []
 
     func start() -> UIViewController {
-        let tabBarController = rootViewController as? UITabBarController
-
         let homeVc = HomeViewController()
         let exploreVc = ExploreViewController()
         let checkinVc = CheckinViewController()
-        tabBarController?.viewControllers = [homeVc, exploreVc, checkinVc]
-
-        let font = UIFont.openSans(.semibold, size: 10)
-
-        let appearance = UITabBarAppearance()
-        appearance.backgroundColor = .white
-        appearance.shadowColor = .secondaryLabel
-
-        appearance.stackedLayoutAppearance.normal.iconColor = .secondaryLabel
-        appearance.stackedLayoutAppearance.normal.titleTextAttributes = [
-            .font: font,
-            .foregroundColor: UIColor.secondaryLabel
-        ]
-
-        appearance.stackedLayoutAppearance.selected.iconColor = .shipGray
-        appearance.stackedLayoutAppearance.selected.titleTextAttributes = [
-            .font: font,
-            .foregroundColor: UIColor.shipGray!
-        ]
-
-        tabBarController?.tabBar.standardAppearance = appearance
+        
+        (rootViewController as? BaseTabBarController)?.viewControllers = [homeVc, exploreVc, checkinVc]
 
         return rootViewController!
     }
